Refactor ProjectsList so that its reusable
